+++ internal/platform/platform.go
@@ -48,6 +48,8 @@ type Provider struct {
 	Init       func(f *resource.Fetcher) error
 	Status     func(stageName string, f resource.Fetcher, e error) error
 	DelConfig  func(f *resource.Fetcher) error
+	// Generates an azure-specific Ignition config.
+	GenerateCloudConfig func(f *resource.Fetcher) (types.Config, error)
 
 	// Fetch, and also save output files to be written during files stage.
 	// Avoid, unless you're certain you need it.
@@ -102,6 +104,13 @@ func (c Config) DelConfig(f *resource.Fetcher) error {
 	} else {
 		return ErrCannotDelete
 	}
+}
+
+func (c Config) GenerateConfig(f *resource.Fetcher) (types.Config, error) {
+	if c.p.GenerateCloudConfig != nil {
+		return c.p.GenerateCloudConfig(f)
+	}
+	return types.Config{}, ErrNoProvider
 }
 
 var configs = registry.Create("platform configs")